@@ -24,19 +24,12 @@ mod plain_tests {
     use http::StatusCode;
     use lychee_lib::Redirects;
     use lychee_lib::{ErrorKind, Status, Uri};
-
-    // Helper function to create a ResponseBody with a given status and URI
-    fn mock_response_body(status: Status, uri: &str) -> ResponseBody {
-        ResponseBody {
-            uri: Uri::try_from(uri).unwrap(),
-            status,
-        }
-    }
+    use test_utils::mock_response_body;
 
     #[test]
     fn test_format_response_with_ok_status() {
         let formatter = PlainFormatter;
-        let body = mock_response_body(Status::Ok(StatusCode::OK), "https://example.com");
+        let body = mock_response_body!(Status::Ok(StatusCode::OK), "https://example.com");
         assert_eq!(
             formatter.format_response(&body),
             "[200] https://example.com/"
@@ -46,20 +39,20 @@ mod plain_tests {
     #[test]
     fn test_format_response_with_error_status() {
         let formatter = PlainFormatter;
-        let body = mock_response_body(
-            Status::Error(ErrorKind::TestError),
+        let body = mock_response_body!(
+            Status::Error(ErrorKind::EmptyUrl),
             "https://example.com/404",
         );
         assert_eq!(
             formatter.format_response(&body),
-            "[ERROR] https://example.com/404 | Generic test error: Test error for formatter testing"
+            "[ERROR] https://example.com/404 | URL cannot be empty: Empty URL found. Check for missing links or malformed markdown"
         );
     }
 
     #[test]
     fn test_format_response_with_excluded_status() {
         let formatter = PlainFormatter;
-        let body = mock_response_body(Status::Excluded, "https://example.com/not-checked");
+        let body = mock_response_body!(Status::Excluded, "https://example.com/not-checked");
         assert_eq!(
             formatter.format_response(&body),
             "[EXCLUDED] https://example.com/not-checked"
@@ -69,7 +62,7 @@ mod plain_tests {
     #[test]
     fn test_format_response_with_redirect_status() {
         let formatter = PlainFormatter;
-        let body = mock_response_body(
+        let body = mock_response_body!(
             Status::Redirected(
                 StatusCode::MOVED_PERMANENTLY,
                 Redirects::from(vec![
@@ -88,7 +81,7 @@ mod plain_tests {
     #[test]
     fn test_format_response_with_unknown_status_code() {
         let formatter = PlainFormatter;
-        let body = mock_response_body(
+        let body = mock_response_body!(
             Status::UnknownStatusCode(StatusCode::from_u16(999).unwrap()),
             "https://example.com/unknown",
         );
