@@ -6,7 +6,7 @@ mod cli {
     use http::{Method, StatusCode};
     use lychee_lib::{InputSource, ResponseBody};
     use predicates::{
-        prelude::{PredicateBooleanExt, predicate},
+        prelude::PredicateBooleanExt,
         str::{contains, is_empty},
     };
     use pretty_assertions::assert_eq;
@@ -18,11 +18,13 @@ mod cli {
         error::Error,
         fs::{self, File},
         io::{BufRead, Write},
-        path::{Path, PathBuf},
+        path::Path,
         time::Duration,
     };
-    use tempfile::NamedTempFile;
-    use test_utils::{mock_server, redirecting_mock_server};
+    use tempfile::{NamedTempFile, tempdir};
+    use test_utils::{
+        fixtures_path, main_command, mock_server, redirecting_mock_server, root_path,
+    };
 
     use uuid::Uuid;
     use wiremock::{
@@ -50,24 +52,6 @@ mod cli {
         }};
     }
 
-    /// Gets the "main" binary name (e.g. `lychee`)
-    fn main_command() -> Command {
-        Command::cargo_bin(env!("CARGO_PKG_NAME")).expect("Couldn't get cargo package name")
-    }
-
-    /// Get the root path of the project.
-    fn root_path() -> PathBuf {
-        Path::new(env!("CARGO_MANIFEST_DIR"))
-            .parent()
-            .unwrap()
-            .to_path_buf()
-    }
-
-    /// Get the path to the fixtures directory.
-    fn fixtures_path() -> PathBuf {
-        root_path().join("fixtures")
-    }
-
     /// Convert a relative path to an absolute path string
     /// starting from a base directory.
     fn path_str(base: &Path, relative_path: &str) -> String {
@@ -77,7 +61,7 @@ mod cli {
     /// Assert actual output lines equals to expected lines.
     /// Order of the lines is ignored.
     fn assert_lines_eq<S: AsRef<str> + Ord>(result: Assert, mut expected_lines: Vec<S>) {
-        let output = result.get_output().stdout.clone();
+        let output = &result.get_output().stdout;
         let mut actual_lines: Vec<String> = output
             .lines()
             .map(|line| line.unwrap().to_string())
@@ -97,8 +81,8 @@ mod cli {
     /// Test the output of the JSON format.
     macro_rules! test_json_output {
         ($test_file:expr, $expected:expr $(, $arg:expr)*) => {{
-            let mut cmd = main_command();
-            let test_path = fixtures_path().join($test_file);
+            let mut cmd = main_command!();
+            let test_path = fixtures_path!().join($test_file);
             let outfile = format!("{}.json", uuid::Uuid::new_v4());
 
             let result = cmd$(.arg($arg))*.arg("--output").arg(&outfile).arg("--format").arg("json").arg(test_path).assert();
@@ -138,9 +122,9 @@ mod cli {
     /// sure that the status code only occurs once.
     #[test]
     fn test_compact_output_format_contains_status() -> Result<()> {
-        let test_path = fixtures_path().join("TEST_INVALID_URLS.html");
+        let test_path = fixtures_path!().join("TEST_INVALID_URLS.html");
 
-        let mut cmd = main_command();
+        let mut cmd = main_command!();
         cmd.arg("--format")
             .arg("compact")
             .arg("--mode")
@@ -182,7 +166,7 @@ mod cli {
     async fn test_json_output() -> Result<()> {
         // Server that returns a bunch of 200 OK responses
         let mock_server_ok = mock_server!(StatusCode::OK);
-        let mut cmd = main_command();
+        let mut cmd = main_command!();
         cmd.arg("--format")
             .arg("json")
             .arg("-vv")
@@ -237,9 +221,9 @@ mod cli {
     /// See https://github.com/lycheeverse/lychee/issues/1355
     #[test]
     fn test_valid_json_output_to_stdout_on_error() -> Result<()> {
-        let test_path = fixtures_path().join("TEST_GITHUB_404.md");
+        let test_path = fixtures_path!().join("TEST_GITHUB_404.md");
 
-        let mut cmd = main_command();
+        let mut cmd = main_command!();
         cmd.arg("--format")
             .arg("json")
             .arg(test_path)
@@ -256,9 +240,9 @@ mod cli {
 
     #[test]
     fn test_detailed_json_output_on_error() -> Result<()> {
-        let test_path = fixtures_path().join("TEST_DETAILED_JSON_OUTPUT_ERROR.md");
+        let test_path = fixtures_path!().join("TEST_DETAILED_JSON_OUTPUT_ERROR.md");
 
-        let mut cmd = main_command();
+        let mut cmd = main_command!();
         cmd.arg("--format")
             .arg("json")
             .arg(&test_path)
@@ -337,10 +321,10 @@ mod cli {
 
     #[test]
     fn test_email_html_with_subject() -> Result<()> {
-        let mut cmd = main_command();
-        let input = fixtures_path().join("TEST_EMAIL_QUERY_PARAMS.html");
+        let input = fixtures_path!().join("TEST_EMAIL_QUERY_PARAMS.html");
 
-        cmd.arg("--dump")
+        main_command!()
+            .arg("--dump")
             .arg(input)
             .arg("--include-mail")
             .assert()
@@ -352,10 +336,10 @@ mod cli {
 
     #[test]
     fn test_email_markdown_with_subject() -> Result<()> {
-        let mut cmd = main_command();
-        let input = fixtures_path().join("TEST_EMAIL_QUERY_PARAMS.md");
+        let input = fixtures_path!().join("TEST_EMAIL_QUERY_PARAMS.md");
 
-        cmd.arg("--dump")
+        main_command!()
+            .arg("--dump")
             .arg(input)
             .arg("--include-mail")
             .assert()
@@ -392,13 +376,13 @@ mod cli {
     /// Test unsupported URI schemes
     #[test]
     fn test_unsupported_uri_schemes_are_ignored() {
-        let mut cmd = main_command();
-        let test_schemes_path = fixtures_path().join("TEST_SCHEMES.txt");
+        let test_schemes_path = fixtures_path!().join("TEST_SCHEMES.txt");
 
         // Exclude file link because it doesn't exist on the filesystem.
         // (File URIs are absolute paths, which we don't have.)
         // Nevertheless, the `file` scheme should be recognized.
-        cmd.arg(test_schemes_path)
+        main_command!()
+            .arg(test_schemes_path)
             .arg("--exclude")
             .arg("file://")
             .env_clear()
@@ -411,10 +395,10 @@ mod cli {
 
     #[test]
     fn test_resolve_paths() {
-        let mut cmd = main_command();
-        let dir = fixtures_path().join("resolve_paths");
+        let dir = fixtures_path!().join("resolve_paths");
 
-        cmd.arg("--offline")
+        main_command!()
+            .arg("--offline")
             .arg("--root-dir")
             .arg(&dir)
             .arg(dir.join("index.html"))
@@ -427,10 +411,10 @@ mod cli {
 
     #[test]
     fn test_resolve_paths_from_root_dir() {
-        let mut cmd = main_command();
-        let dir = fixtures_path().join("resolve_paths_from_root_dir");
+        let dir = fixtures_path!().join("resolve_paths_from_root_dir");
 
-        cmd.arg("--offline")
+        main_command!()
+            .arg("--offline")
             .arg("--include-fragments")
             .arg("--root-dir")
             .arg(&dir)
@@ -445,10 +429,10 @@ mod cli {
 
     #[test]
     fn test_resolve_paths_from_root_dir_and_base_url() {
-        let mut cmd = main_command();
-        let dir = fixtures_path();
+        let dir = fixtures_path!();
 
-        cmd.arg("--offline")
+        main_command!()
+            .arg("--offline")
             .arg("--root-dir")
             .arg(dir.join("resolve_paths"))
             .arg("--base-url")
@@ -465,7 +449,7 @@ mod cli {
     fn test_youtube_quirk() {
         let url = "https://www.youtube.com/watch?v=NlKuICiT470&list=PLbWDhxwM_45mPVToqaIZNbZeIzFchsKKQ&index=7";
 
-        main_command()
+        main_command!()
             .write_stdin(url)
             .arg("--verbose")
             .arg("--no-progress")
@@ -480,7 +464,7 @@ mod cli {
     fn test_crates_io_quirk() {
         let url = "https://crates.io/crates/lychee";
 
-        main_command()
+        main_command!()
             .write_stdin(url)
             .arg("--verbose")
             .arg("--no-progress")
@@ -498,7 +482,7 @@ mod cli {
     fn test_ignored_hosts() {
         let url = "https://twitter.com/zarfeblong/status/1339742840142872577";
 
-        main_command()
+        main_command!()
             .write_stdin(url)
             .arg("--verbose")
             .arg("--no-progress")
@@ -517,8 +501,8 @@ mod cli {
         let mut file = File::create(&file_path)?;
         writeln!(file, "{}", mock_server.uri())?;
 
-        let mut cmd = main_command();
-        cmd.arg(file_path)
+        main_command!()
+            .arg(file_path)
             .write_stdin(mock_server.uri())
             .assert()
             .failure()
@@ -529,10 +513,10 @@ mod cli {
 
     #[test]
     fn test_schemes() {
-        let mut cmd = main_command();
-        let test_schemes_path = fixtures_path().join("TEST_SCHEMES.md");
+        let test_schemes_path = fixtures_path!().join("TEST_SCHEMES.md");
 
-        cmd.arg(test_schemes_path)
+        main_command!()
+            .arg(test_schemes_path)
             .arg("--scheme")
             .arg("https")
             .arg("--scheme")
@@ -547,11 +531,11 @@ mod cli {
 
     #[test]
     fn test_caching_single_file() {
-        let mut cmd = main_command();
         // Repetitions in one file shall all be checked and counted only once.
-        let test_schemes_path_1 = fixtures_path().join("TEST_REPETITION_1.txt");
+        let test_schemes_path_1 = fixtures_path!().join("TEST_REPETITION_1.txt");
 
-        cmd.arg(&test_schemes_path_1)
+        main_command!()
+            .arg(&test_schemes_path_1)
             .env_clear()
             .assert()
             .success()
@@ -563,7 +547,7 @@ mod cli {
     // Test that two identical requests don't get executed twice.
     fn test_caching_across_files() -> Result<()> {
         // Repetitions across multiple files shall all be checked only once.
-        let repeated_uris = fixtures_path().join("TEST_REPETITION_*.txt");
+        let repeated_uris = fixtures_path!().join("TEST_REPETITION_*.txt");
 
         test_json_output!(
             repeated_uris,
@@ -584,10 +568,10 @@ mod cli {
 
     #[test]
     fn test_failure_github_404_no_token() {
-        let mut cmd = main_command();
-        let test_github_404_path = fixtures_path().join("TEST_GITHUB_404.md");
+        let test_github_404_path = fixtures_path!().join("TEST_GITHUB_404.md");
 
-        cmd.arg(test_github_404_path)
+        main_command!()
+            .arg(test_github_404_path)
             .arg("--no-progress")
             .env_clear()
             .assert()
@@ -603,10 +587,10 @@ mod cli {
 
     #[tokio::test]
     async fn test_stdin_input() {
-        let mut cmd = main_command();
         let mock_server = mock_server!(StatusCode::OK);
 
-        cmd.arg("-")
+        main_command!()
+            .arg("-")
             .write_stdin(mock_server.uri())
             .assert()
             .success();
@@ -614,10 +598,10 @@ mod cli {
 
     #[tokio::test]
     async fn test_stdin_input_failure() {
-        let mut cmd = main_command();
         let mock_server = mock_server!(StatusCode::INTERNAL_SERVER_ERROR);
 
-        cmd.arg("-")
+        main_command!()
+            .arg("-")
             .write_stdin(mock_server.uri())
             .assert()
             .failure()
@@ -626,13 +610,13 @@ mod cli {
 
     #[tokio::test]
     async fn test_stdin_input_multiple() {
-        let mut cmd = main_command();
         let mock_server_a = mock_server!(StatusCode::OK);
         let mock_server_b = mock_server!(StatusCode::OK);
 
         // this behavior (treating multiple `-` as separate inputs) is the same as most CLI tools
         // that accept `-` as stdin, e.g. `cat`, `bat`, `grep` etc.
-        cmd.arg("-")
+        main_command!()
+            .arg("-")
             .arg("-")
             .write_stdin(mock_server_a.uri())
             .write_stdin(mock_server_b.uri())
@@ -642,55 +626,114 @@ mod cli {
 
     #[test]
     fn test_missing_file_ok_if_skip_missing() {
-        let mut cmd = main_command();
         let filename = format!("non-existing-file-{}", uuid::Uuid::new_v4());
-
-        cmd.arg(&filename).arg("--skip-missing").assert().success();
+        main_command!()
+            .arg(&filename)
+            .arg("--skip-missing")
+            .assert()
+            .success();
     }
 
     #[test]
     fn test_skips_hidden_files_by_default() {
-        main_command()
-            .arg(fixtures_path().join("hidden/"))
+        main_command!()
+            .arg(fixtures_path!().join("hidden/"))
             .assert()
             .success()
             .stdout(contains("0 Total"));
+
+        main_command!()
+            .arg("--dump")
+            .arg(fixtures_path!().join("hidden/"))
+            .assert()
+            .stdout("")
+            .success();
+
+        main_command!()
+            .arg("--dump-inputs")
+            .arg(fixtures_path!().join("hidden/"))
+            .assert()
+            .stdout("")
+            .success();
     }
 
     #[test]
     fn test_include_hidden_file() {
-        main_command()
-            .arg(fixtures_path().join("hidden/"))
+        main_command!()
+            .arg(fixtures_path!().join("hidden/"))
             .arg("--hidden")
             .assert()
             .success()
             .stdout(contains("1 Total"));
+
+        main_command!()
+            .arg("--dump")
+            .arg("--hidden")
+            .arg(fixtures_path!().join("hidden/"))
+            .assert()
+            .stdout(contains("wikipedia.org"))
+            .success();
+
+        main_command!()
+            .arg("--dump-inputs")
+            .arg("--hidden")
+            .arg(fixtures_path!().join("hidden/"))
+            .assert()
+            .stdout(contains(".hidden"))
+            .success();
     }
 
     #[test]
     fn test_skips_ignored_files_by_default() {
-        main_command()
-            .arg(fixtures_path().join("ignore/"))
+        main_command!()
+            .arg(fixtures_path!().join("ignore/"))
             .assert()
             .success()
             .stdout(contains("0 Total"));
+
+        main_command!()
+            .arg("--dump")
+            .arg(fixtures_path!().join("ignore/"))
+            .assert()
+            .success()
+            .stdout("");
+
+        main_command!()
+            .arg("--dump-inputs")
+            .arg(fixtures_path!().join("ignore/"))
+            .assert()
+            .success()
+            .stdout("");
     }
 
     #[test]
     fn test_include_ignored_file() {
-        main_command()
-            .arg(fixtures_path().join("ignore/"))
+        main_command!()
+            .arg(fixtures_path!().join("ignore/"))
             .arg("--no-ignore")
             .assert()
             .success()
             .stdout(contains("1 Total"));
+
+        main_command!()
+            .arg("--dump")
+            .arg("--no-ignore")
+            .arg(fixtures_path!().join("ignore/"))
+            .assert()
+            .success()
+            .stdout(contains("wikipedia.org"));
+
+        main_command!()
+            .arg("--dump-inputs")
+            .arg("--no-ignore")
+            .arg(fixtures_path!().join("ignore/"))
+            .assert()
+            .success()
+            .stdout(contains("ignored-file.md"));
     }
 
     #[tokio::test]
     async fn test_glob() -> Result<()> {
-        // using Result to be able to use `?`
-        let mut cmd = main_command();
-
         let dir = tempfile::tempdir()?;
         let mock_server_a = mock_server!(StatusCode::OK);
         let mock_server_b = mock_server!(StatusCode::OK);
@@ -700,7 +743,8 @@ mod cli {
         writeln!(file_a, "{}", mock_server_a.uri().as_str())?;
         writeln!(file_b, "{}", mock_server_b.uri().as_str())?;
 
-        cmd.arg(dir.path().join("*.md"))
+        main_command!()
+            .arg(dir.path().join("*.md"))
             .arg("--verbose")
             .assert()
             .success()
@@ -712,8 +756,6 @@ mod cli {
     #[cfg(target_os = "linux")] // MacOS and Windows have case-insensitive filesystems
     #[tokio::test]
     async fn test_glob_ignore_case() -> Result<()> {
-        let mut cmd = main_command();
-
         let dir = tempfile::tempdir()?;
         let mock_server_a = mock_server!(StatusCode::OK);
         let mock_server_b = mock_server!(StatusCode::OK);
@@ -723,7 +765,8 @@ mod cli {
         writeln!(file_a, "{}", mock_server_a.uri().as_str())?;
         writeln!(file_b, "{}", mock_server_b.uri().as_str())?;
 
-        cmd.arg(dir.path().join("[r]eadme.md"))
+        main_command!()
+            .arg(dir.path().join("[r]eadme.md"))
             .arg("--verbose")
             .arg("--glob-ignore-case")
             .assert()
@@ -735,8 +778,6 @@ mod cli {
 
     #[tokio::test]
     async fn test_glob_recursive() -> Result<()> {
-        let mut cmd = main_command();
-
         let dir = tempfile::tempdir()?;
         let subdir_level_1 = tempfile::tempdir_in(&dir)?;
         let subdir_level_2 = tempfile::tempdir_in(&subdir_level_1)?;
@@ -746,8 +787,8 @@ mod cli {
 
         writeln!(file, "{}", mock_server.uri().as_str())?;
 
-        // ** should be a recursive glob
-        cmd.arg(dir.path().join("**/*.md"))
+        main_command!()
+            .arg(dir.path().join("**/*.md")) // ** should be a recursive glob
             .arg("--verbose")
             .assert()
             .success()
@@ -773,11 +814,11 @@ mod cli {
     /// Test writing output of `--dump` command to file
     #[test]
     fn test_dump_to_file() -> Result<()> {
-        let mut cmd = main_command();
-        let test_path = fixtures_path().join("TEST.md");
+        let test_path = fixtures_path!().join("TEST.md");
         let outfile = format!("{}", Uuid::new_v4());
 
-        cmd.arg("--output")
+        main_command!()
+            .arg("--output")
             .arg(&outfile)
             .arg("--dump")
             .arg("--include-mail")
@@ -799,10 +840,10 @@ mod cli {
     /// Test excludes
     #[test]
     fn test_exclude_wildcard() -> Result<()> {
-        let mut cmd = main_command();
-        let test_path = fixtures_path().join("TEST.md");
+        let test_path = fixtures_path!().join("TEST.md");
 
-        cmd.arg(test_path)
+        main_command!()
+            .arg(test_path)
             .arg("--exclude")
             .arg(".*")
             .assert()
@@ -814,10 +855,10 @@ mod cli {
 
     #[test]
     fn test_exclude_multiple_urls() -> Result<()> {
-        let mut cmd = main_command();
-        let test_path = fixtures_path().join("TEST.md");
+        let test_path = fixtures_path!().join("TEST.md");
 
-        cmd.arg(test_path)
+        main_command!()
+            .arg(test_path)
             .arg("--exclude")
             .arg("https://en.wikipedia.org/*")
             .arg("--exclude")
@@ -830,11 +871,11 @@ mod cli {
     }
 
     #[tokio::test]
-    async fn test_empty_config() -> Result<()> {
+    async fn test_empty_config() {
         let mock_server = mock_server!(StatusCode::OK);
-        let config = fixtures_path().join("configs").join("empty.toml");
-        let mut cmd = main_command();
-        cmd.arg("--config")
+        let config = fixtures_path!().join("configs").join("empty.toml");
+        main_command!()
+            .arg("--config")
             .arg(config)
             .arg("-")
             .write_stdin(mock_server.uri())
@@ -843,21 +884,17 @@ mod cli {
             .success()
             .stdout(contains("1 Total"))
             .stdout(contains("1 OK"));
-
-        Ok(())
     }
 
     #[test]
-    fn test_invalid_default_config() -> Result<()> {
-        let test_path = fixtures_path().join("configs");
-        let mut cmd = main_command();
+    fn test_invalid_default_config() {
+        let test_path = fixtures_path!().join("configs");
+        let mut cmd = main_command!();
         cmd.current_dir(test_path)
             .arg(".")
             .assert()
             .failure()
             .stderr(contains("Cannot load default configuration file"));
-
-        Ok(())
     }
 
     #[tokio::test]
@@ -867,8 +904,8 @@ mod cli {
         let mut config = NamedTempFile::new()?;
         writeln!(config, "include_mail = false")?;
 
-        let mut cmd = main_command();
-        cmd.arg("--config")
+        main_command!()
+            .arg("--config")
             .arg(config.path().to_str().unwrap())
             .arg("-")
             .write_stdin(test_mail_address)
@@ -881,8 +918,8 @@ mod cli {
         let mut config = NamedTempFile::new()?;
         writeln!(config, "include_mail = true")?;
 
-        let mut cmd = main_command();
-        cmd.arg("--config")
+        main_command!()
+            .arg("--config")
             .arg(config.path().to_str().unwrap())
             .arg("-")
             .write_stdin(test_mail_address)
@@ -898,9 +935,9 @@ mod cli {
     #[tokio::test]
     async fn test_cache_config() -> Result<()> {
         let mock_server = mock_server!(StatusCode::OK);
-        let config = fixtures_path().join("configs").join("cache.toml");
-        let mut cmd = main_command();
-        cmd.arg("--config")
+        let config = fixtures_path!().join("configs").join("cache.toml");
+        main_command!()
+            .arg("--config")
             .arg(config)
             .arg("-")
             .write_stdin(mock_server.uri())
@@ -915,24 +952,40 @@ mod cli {
 
     #[tokio::test]
     async fn test_invalid_config() {
-        let config = fixtures_path().join("configs").join("invalid.toml");
-        let mut cmd = main_command();
-        cmd.arg("--config")
+        let config = fixtures_path!().join("configs").join("invalid.toml");
+        main_command!()
+            .arg("--config")
             .arg(config)
             .arg("-")
             .env_clear()
             .assert()
             .failure()
-            .stderr(predicate::str::contains("Cannot load configuration file"))
-            .stderr(predicate::str::contains("Failed to parse"))
-            .stderr(predicate::str::contains("TOML parse error"));
+            .stderr(contains("Cannot load configuration file"))
+            .stderr(contains("Failed to parse"))
+            .stderr(contains("TOML parse error"));
+    }
+
+    #[tokio::test]
+    async fn test_config_invalid_keys() {
+        let mock_server = mock_server!(StatusCode::OK);
+        let config = fixtures_path!().join("configs").join("invalid-key.toml");
+        main_command!()
+            .arg("--config")
+            .arg(config)
+            .arg("-")
+            .write_stdin(mock_server.uri())
+            .env_clear()
+            .assert()
+            .failure()
+            .code(3)
+            .stderr(contains("unknown field `this_is_invalid`, expected one of"));
     }
 
     #[tokio::test]
     async fn test_missing_config_error() {
         let mock_server = mock_server!(StatusCode::OK);
-        let mut cmd = main_command();
-        cmd.arg("--config")
+        main_command!()
+            .arg("--config")
             .arg("config.does.not.exist.toml")
             .arg("-")
             .write_stdin(mock_server.uri())
@@ -944,9 +997,9 @@ mod cli {
     #[tokio::test]
     async fn test_config_example() {
         let mock_server = mock_server!(StatusCode::OK);
-        let config = root_path().join("lychee.example.toml");
-        let mut cmd = main_command();
-        cmd.arg("--config")
+        let config = root_path!().join("lychee.example.toml");
+        main_command!()
+            .arg("--config")
             .arg(config)
             .arg("-")
             .write_stdin(mock_server.uri())
@@ -955,12 +1008,53 @@ mod cli {
             .success();
     }
 
+    #[test]
+    #[cfg(unix)]
+    fn test_all_arguments_in_config() -> Result<()> {
+        let help_cmd = main_command!().env_clear().arg("--help").assert().success();
+        let help_text = std::str::from_utf8(&help_cmd.get_output().stdout)?;
+
+        let regex = test_utils::arg_regex_help!()?;
+        let excluded = [
+            "base",         // deprecated
+            "exclude_file", // deprecated
+            "config",       // not part of config
+            "quiet",        // not part of config
+            "help",         // special clap argument
+            "version",      // special clap argument
+        ];
+
+        let arguments: Vec<String> = help_text
+            .lines()
+            .filter_map(|line| {
+                let captures = regex.captures(line)?;
+                captures.name("long").map(|m| m.as_str())
+            })
+            .map(|arg| arg.replace("-", "_"))
+            .filter(|arg| !excluded.contains(&arg.as_str()))
+            .collect();
+
+        let config = root_path!().join("lychee.example.toml");
+        let values: toml::Table = toml::from_str(&std::fs::read_to_string(config)?)?;
+
+        for argument in arguments {
+            if !values.contains_key(&argument) {
+                panic!(
+                    "Key '{argument}' missing in config.
+The config file should contain every possible key for documentation purposes."
+                )
+            }
+        }
+
+        Ok(())
+    }
+
     #[tokio::test]
     async fn test_config_smoketest() {
         let mock_server = mock_server!(StatusCode::OK);
-        let config = fixtures_path().join("configs").join("smoketest.toml");
-        let mut cmd = main_command();
-        cmd.arg("--config")
+        let config = fixtures_path!().join("configs").join("smoketest.toml");
+        main_command!()
+            .arg("--config")
             .arg(config)
             .arg("-")
             .write_stdin(mock_server.uri())
@@ -972,9 +1066,9 @@ mod cli {
     #[tokio::test]
     async fn test_config_accept() {
         let mock_server = mock_server!(StatusCode::OK);
-        let config = fixtures_path().join("configs").join("accept.toml");
-        let mut cmd = main_command();
-        cmd.arg("--config")
+        let config = fixtures_path!().join("configs").join("accept.toml");
+        main_command!()
+            .arg("--config")
             .arg(config)
             .arg("-")
             .write_stdin(mock_server.uri())
@@ -983,12 +1077,25 @@ mod cli {
             .success();
     }
 
+    #[tokio::test]
+    #[cfg(unix)]
+    async fn test_config_files_from() {
+        let dir = fixtures_path!().join("configs").join("files_from");
+        let result = main_command!()
+            .current_dir(dir)
+            .arg("/dev/null") // at least one input arg is required. this could be changed in the future
+            .arg("--dump")
+            .assert()
+            .success();
+
+        assert_lines_eq(result, vec!["https://wikipedia.org/"]);
+    }
+
     #[test]
     fn test_lycheeignore_file() -> Result<()> {
-        let mut cmd = main_command();
-        let test_path = fixtures_path().join("lycheeignore");
+        let test_path = fixtures_path!().join("lycheeignore");
 
-        let cmd = cmd
+        let cmd = main_command!()
             .current_dir(test_path)
             .arg("--dump")
             .arg("TEST.md")
@@ -1006,11 +1113,11 @@ mod cli {
 
     #[test]
     fn test_lycheeignore_and_exclude_file() -> Result<()> {
-        let mut cmd = main_command();
-        let test_path = fixtures_path().join("lycheeignore");
+        let test_path = fixtures_path!().join("lycheeignore");
         let excludes_path = test_path.join("normal-exclude-file");
 
-        cmd.current_dir(test_path)
+        main_command!()
+            .current_dir(test_path)
             .arg("TEST.md")
             .arg("--exclude-file")
             .arg(excludes_path)
@@ -1024,7 +1131,7 @@ mod cli {
 
     #[tokio::test]
     async fn test_lycheecache_file() -> Result<()> {
-        let base_path = fixtures_path().join("cache");
+        let base_path = fixtures_path!().join("cache");
         let cache_file = base_path.join(LYCHEE_CACHE_FILE);
 
         // Ensure clean state
@@ -1049,7 +1156,7 @@ mod cli {
         file.sync_all()?;
 
         // Create and run command
-        let mut cmd = main_command();
+        let mut cmd = main_command!();
         cmd.current_dir(&base_path)
             .arg(&file_path)
             .arg("--verbose")
@@ -1096,11 +1203,7 @@ mod cli {
 
         // Clean up
         fs::remove_file(&cache_file).map_err(|e| {
-            anyhow::anyhow!(
-                "Failed to remove cache file: {:?}, error: {}",
-                cache_file,
-                e
-            )
+            anyhow::anyhow!("Failed to remove cache file: {cache_file:?}, error: {e}")
         })?;
 
         Ok(())
@@ -1108,7 +1211,7 @@ mod cli {
 
     #[tokio::test]
     async fn test_lycheecache_exclude_custom_status_codes() -> Result<()> {
-        let base_path = fixtures_path().join("cache");
+        let base_path = fixtures_path!().join("cache");
         let cache_file = base_path.join(LYCHEE_CACHE_FILE);
 
         // Unconditionally remove cache file if it exists
@@ -1125,7 +1228,7 @@ mod cli {
         writeln!(file, "{}", mock_server_no_content.uri().as_str())?;
         writeln!(file, "{}", mock_server_too_many_requests.uri().as_str())?;
 
-        let mut cmd = main_command();
+        let mut cmd = main_command!();
         let test_cmd = cmd
             .current_dir(&base_path)
             .arg(dir.path().join("c.md"))
@@ -1171,7 +1274,7 @@ mod cli {
 
     #[tokio::test]
     async fn test_lycheecache_accept_custom_status_codes() -> Result<()> {
-        let base_path = fixtures_path().join("cache_accept_custom_status_codes");
+        let base_path = fixtures_path!().join("cache_accept_custom_status_codes");
         let cache_file = base_path.join(LYCHEE_CACHE_FILE);
 
         // Unconditionally remove cache file if it exists
@@ -1188,7 +1291,7 @@ mod cli {
         writeln!(file, "{}", mock_server_teapot.uri().as_str())?;
         writeln!(file, "{}", mock_server_server_error.uri().as_str())?;
 
-        let mut cmd = main_command();
+        let mut cmd = main_command!();
         let test_cmd = cmd
             .current_dir(&base_path)
             .arg(dir.path().join("c.md"))
@@ -1248,8 +1351,8 @@ mod cli {
     async fn test_accept_overrides_defaults_not_additive() -> Result<()> {
         let mock_server_200 = mock_server!(StatusCode::OK);
 
-        let mut cmd = main_command();
-        cmd.arg("--accept")
+        main_command!()
+            .arg("--accept")
             .arg("404") // ONLY accept 404 - should reject 200 as we overwrite the default
             .arg("-")
             .write_stdin(mock_server_200.uri())
@@ -1266,7 +1369,7 @@ mod cli {
 
     #[tokio::test]
     async fn test_skip_cache_unsupported() -> Result<()> {
-        let base_path = fixtures_path().join("cache");
+        let base_path = fixtures_path!().join("cache");
         let cache_file = base_path.join(LYCHEE_CACHE_FILE);
 
         // Unconditionally remove cache file if it exists
@@ -1276,7 +1379,7 @@ mod cli {
         let excluded_url = "https://example.com/";
 
         // run first without cache to generate the cache file
-        main_command()
+        main_command!()
             .current_dir(&base_path)
             .write_stdin(format!("{unsupported_url}\n{excluded_url}"))
             .arg("--cache")
@@ -1317,7 +1420,7 @@ mod cli {
     /// status codes.
     #[tokio::test]
     async fn test_skip_cache_unknown_status_code() -> Result<()> {
-        let base_path = fixtures_path().join("cache");
+        let base_path = fixtures_path!().join("cache");
         let cache_file = base_path.join(LYCHEE_CACHE_FILE);
 
         // Unconditionally remove cache file if it exists
@@ -1328,7 +1431,7 @@ mod cli {
         let unknown_url = "https://www.linkedin.com/company/corrode";
 
         // run first without cache to generate the cache file
-        main_command()
+        main_command!()
             .current_dir(&base_path)
             .write_stdin(unknown_url.to_string())
             .arg("--cache")
@@ -1359,25 +1462,23 @@ mod cli {
     }
 
     #[test]
-    fn test_verbatim_skipped_by_default() -> Result<()> {
-        let mut cmd = main_command();
-        let input = fixtures_path().join("TEST_CODE_BLOCKS.md");
+    fn test_verbatim_skipped_by_default() {
+        let input = fixtures_path!().join("TEST_CODE_BLOCKS.md");
 
-        cmd.arg(input)
+        main_command!()
+            .arg(input)
             .arg("--dump")
             .assert()
             .success()
             .stdout(is_empty());
-
-        Ok(())
     }
 
     #[test]
-    fn test_include_verbatim() -> Result<()> {
-        let mut cmd = main_command();
-        let input = fixtures_path().join("TEST_CODE_BLOCKS.md");
+    fn test_include_verbatim() {
+        let input = fixtures_path!().join("TEST_CODE_BLOCKS.md");
 
-        cmd.arg("--include-verbatim")
+        main_command!()
+            .arg("--include-verbatim")
             .arg(input)
             .arg("--dump")
             .assert()
@@ -1385,47 +1486,41 @@ mod cli {
             .stdout(contains("http://127.0.0.1/block"))
             .stdout(contains("http://127.0.0.1/inline"))
             .stdout(contains("http://127.0.0.1/bash"));
-
-        Ok(())
     }
     #[tokio::test]
-    async fn test_verbatim_skipped_by_default_via_file() -> Result<()> {
-        let file = fixtures_path().join("TEST_VERBATIM.html");
+    async fn test_verbatim_skipped_by_default_via_file() {
+        let file = fixtures_path!().join("TEST_VERBATIM.html");
 
-        main_command()
+        main_command!()
             .arg("--dump")
             .arg(file)
             .assert()
             .success()
             .stdout(is_empty());
-
-        Ok(())
     }
 
     #[tokio::test]
-    async fn test_verbatim_skipped_by_default_via_remote_url() -> Result<()> {
-        let mut cmd = main_command();
-        let file = fixtures_path().join("TEST_VERBATIM.html");
-        let body = fs::read_to_string(file)?;
+    async fn test_verbatim_skipped_by_default_via_remote_url() {
+        let file = fixtures_path!().join("TEST_VERBATIM.html");
+        let body = fs::read_to_string(file).unwrap();
         let mock_server = mock_response!(body);
 
-        cmd.arg("--dump")
+        main_command!()
+            .arg("--dump")
             .arg(mock_server.uri())
             .assert()
             .success()
             .stdout(is_empty());
-
-        Ok(())
     }
 
     #[tokio::test]
-    async fn test_include_verbatim_via_remote_url() -> Result<()> {
-        let mut cmd = main_command();
-        let file = fixtures_path().join("TEST_VERBATIM.html");
-        let body = fs::read_to_string(file)?;
+    async fn test_include_verbatim_via_remote_url() {
+        let file = fixtures_path!().join("TEST_VERBATIM.html");
+        let body = fs::read_to_string(file).unwrap();
         let mock_server = mock_response!(body);
 
-        cmd.arg("--include-verbatim")
+        main_command!()
+            .arg("--include-verbatim")
             .arg("--dump")
             .arg(mock_server.uri())
             .assert()
@@ -1436,64 +1531,54 @@ mod cli {
             .stdout(contains("http://www.example.com/kbd"))
             .stdout(contains("http://www.example.com/var"))
             .stdout(contains("http://www.example.com/script"));
-        Ok(())
     }
 
     #[test]
-    fn test_require_https() -> Result<()> {
-        let mut cmd = main_command();
-        let test_path = fixtures_path().join("TEST_HTTP.html");
-        cmd.arg(&test_path).assert().success();
+    fn test_require_https() {
+        let test_path = fixtures_path!().join("TEST_HTTP.html");
+        main_command!().arg(&test_path).assert().success();
 
-        let mut cmd = main_command();
-        cmd.arg("--require-https")
+        main_command!()
+            .arg("--require-https")
             .arg(test_path)
             .assert()
             .failure()
             .stdout(contains("This URI is available in HTTPS protocol, but HTTP is provided. Use 'https://example.com/' instead"));
-
-        Ok(())
     }
 
     /// If `base-dir` is not set, don't throw an error in case we encounter
     /// an absolute local link (e.g. `/about`) within a file.
     /// Instead, simply ignore the link.
     #[test]
-    fn test_ignore_absolute_local_links_without_base() -> Result<()> {
-        let mut cmd = main_command();
+    fn test_ignore_absolute_local_links_without_base() {
+        let offline_dir = fixtures_path!().join("offline");
 
-        let offline_dir = fixtures_path().join("offline");
-
-        cmd.arg("--offline")
+        main_command!()
+            .arg("--offline")
             .arg(offline_dir.join("index.html"))
             .env_clear()
             .assert()
             .success()
             .stdout(contains("0 Total"));
-
-        Ok(())
     }
 
     #[test]
-    fn test_inputs_without_scheme() -> Result<()> {
-        let test_path = fixtures_path().join("TEST_HTTP.html");
-        let mut cmd = main_command();
-
-        cmd.arg("--dump")
+    fn test_inputs_without_scheme() {
+        let test_path = fixtures_path!().join("TEST_HTTP.html");
+        main_command!()
+            .arg("--dump")
             .arg("example.com")
             .arg(&test_path)
             .arg("https://example.org")
             .assert()
             .success();
-        Ok(())
     }
 
     #[test]
-    fn test_print_excluded_links_in_verbose_mode() -> Result<()> {
-        let test_path = fixtures_path().join("TEST_DUMP_EXCLUDE.txt");
-        let mut cmd = main_command();
-
-        cmd.arg("--dump")
+    fn test_print_excluded_links_in_verbose_mode() {
+        let test_path = fixtures_path!().join("TEST_DUMP_EXCLUDE.txt");
+        main_command!()
+            .arg("--dump")
             .arg("--verbose")
             .arg("--exclude")
             .arg("example.com")
@@ -1513,14 +1598,12 @@ mod cli {
                 "https://example.com/foo/bar ({}) [excluded]",
                 test_path.display()
             )));
-        Ok(())
     }
 
     #[test]
-    fn test_remap_uri() -> Result<()> {
-        let mut cmd = main_command();
-
-        cmd.arg("--dump")
+    fn test_remap_uri() {
+        main_command!()
+            .arg("--dump")
             .arg("--remap")
             .arg("https://example.com http://127.0.0.1:8080")
             .arg("--remap")
@@ -1534,16 +1617,13 @@ mod cli {
             .stdout(contains("http://127.0.0.1:8080/"))
             .stdout(contains("https://staging.example.com/"))
             .stdout(contains("https://example.net/"));
-
-        Ok(())
     }
 
     #[test]
     #[ignore = "Skipping test until https://github.com/robinst/linkify/pull/58 is merged"]
-    fn test_remap_path() -> Result<()> {
-        let mut cmd = main_command();
-
-        cmd.arg("--dump")
+    fn test_remap_path() {
+        main_command!()
+            .arg("--dump")
             .arg("--remap")
             .arg("../../issues https://github.com/usnistgov/OSCAL/issues")
             .arg("--")
@@ -1553,15 +1633,12 @@ mod cli {
             .assert()
             .success()
             .stdout(contains("https://github.com/usnistgov/OSCAL/issues"));
-
-        Ok(())
     }
 
     #[test]
-    fn test_remap_capture() -> Result<()> {
-        let mut cmd = main_command();
-
-        cmd.arg("--dump")
+    fn test_remap_capture() {
+        main_command!()
+            .arg("--dump")
             .arg("--remap")
             .arg("https://example.com/(.*) http://example.org/$1")
             .arg("--")
@@ -1571,15 +1648,12 @@ mod cli {
             .assert()
             .success()
             .stdout(contains("http://example.org/foo"));
-
-        Ok(())
     }
 
     #[test]
-    fn test_remap_named_capture() -> Result<()> {
-        let mut cmd = main_command();
-
-        cmd.arg("--dump")
+    fn test_remap_named_capture() {
+        main_command!()
+            .arg("--dump")
             .arg("--remap")
             .arg("https://github.com/(?P<org>.*)/(?P<repo>.*) https://gitlab.com/$org/$repo")
             .arg("--")
@@ -1589,18 +1663,14 @@ mod cli {
             .assert()
             .success()
             .stdout(contains("https://gitlab.com/lycheeverse/lychee"));
-
-        Ok(())
     }
 
     #[test]
-    fn test_excluded_paths_regex() -> Result<()> {
-        let test_path = fixtures_path().join("exclude-path");
+    fn test_excluded_paths_regex() {
+        let test_path = fixtures_path!().join("exclude-path");
         let excluded_path_1 = "\\/excluded?\\/"; // exclude paths containing a directory "exclude" and "excluded"
         let excluded_path_2 = "(\\.mdx|\\.txt)$"; // exclude .mdx and .txt files
-        let mut cmd = main_command();
-
-        let result = cmd
+        let result = main_command!()
             .arg("--exclude-path")
             .arg(excluded_path_1)
             .arg("--exclude-path")
@@ -1618,16 +1688,14 @@ mod cli {
                 "https://test.md/to-be-included-inner",
             ],
         );
-
-        Ok(())
     }
 
     #[test]
-    fn test_handle_relative_paths_as_input() -> Result<()> {
-        let test_path = fixtures_path();
-        let mut cmd = main_command();
+    fn test_handle_relative_paths_as_input() {
+        let test_path = fixtures_path!();
 
-        cmd.current_dir(&test_path)
+        main_command!()
+            .current_dir(&test_path)
             .arg("--verbose")
             .arg("--exclude")
             .arg("example.*")
@@ -1637,16 +1705,14 @@ mod cli {
             .success()
             .stdout(contains("3 Total"))
             .stdout(contains("3 Excluded"));
-
-        Ok(())
     }
 
     #[test]
-    fn test_handle_nonexistent_relative_paths_as_input() -> Result<()> {
-        let test_path = fixtures_path();
-        let mut cmd = main_command();
+    fn test_handle_nonexistent_relative_paths_as_input() {
+        let test_path = fixtures_path!();
 
-        cmd.current_dir(&test_path)
+        main_command!()
+            .current_dir(&test_path)
             .arg("--verbose")
             .arg("--exclude")
             .arg("example.*")
@@ -1655,23 +1721,19 @@ mod cli {
             .assert()
             .failure()
             .stderr(contains("Invalid file path: ./NOT-A-REAL-TEST-FIXTURE.md"));
-
-        Ok(())
     }
 
     #[test]
-    fn test_prevent_too_many_redirects() -> Result<()> {
-        let mut cmd = main_command();
+    fn test_prevent_too_many_redirects() {
         let url = "https://http.codes/308";
 
-        cmd.write_stdin(url)
+        main_command!()
+            .write_stdin(url)
             .arg("--max-redirects")
             .arg("0")
             .arg("-")
             .assert()
             .failure();
-
-        Ok(())
     }
 
     #[test]
@@ -1681,8 +1743,8 @@ mod cli {
 
         for _ in 0..3 {
             // This can be flaky. Try up to 3 times
-            let mut cmd = main_command();
-            let input = fixtures_path().join("INTERNET_ARCHIVE.md");
+            let mut cmd = main_command!();
+            let input = fixtures_path!().join("INTERNET_ARCHIVE.md");
 
             cmd.arg("--no-progress").arg("--suggest").arg(input);
 
@@ -1708,7 +1770,7 @@ mod cli {
     }
 
     #[tokio::test]
-    async fn test_basic_auth() -> Result<()> {
+    async fn test_basic_auth() {
         let username = "username";
         let password = "password123";
 
@@ -1725,7 +1787,7 @@ mod cli {
             .await;
 
         // Configure the command to use the BasicAuthExtractor
-        main_command()
+        main_command!()
             .arg("--verbose")
             .arg("--basic-auth")
             .arg(format!("{} {username}:{password}", mock_server.uri()))
@@ -1737,7 +1799,7 @@ mod cli {
             .stdout(contains("1 OK"));
 
         // Websites as direct arguments must also use authentication
-        main_command()
+        main_command!()
             .arg(mock_server.uri())
             .arg("--verbose")
             .arg("--basic-auth")
@@ -1745,12 +1807,10 @@ mod cli {
             .assert()
             .success()
             .stdout(contains("0 Total")); // Mock server returns no body, so there are no URLs to check
-
-        Ok(())
     }
 
     #[tokio::test]
-    async fn test_multi_basic_auth() -> Result<()> {
+    async fn test_multi_basic_auth() {
         let username1 = "username";
         let password1 = "password123";
         let mock_server1 = wiremock::MockServer::start().await;
@@ -1769,7 +1829,7 @@ mod cli {
             .await;
 
         // Configure the command to use the BasicAuthExtractor
-        main_command()
+        main_command!()
             .arg("--verbose")
             .arg("--basic-auth")
             .arg(format!("{} {username1}:{password1}", mock_server1.uri()))
@@ -1781,16 +1841,14 @@ mod cli {
             .success()
             .stdout(contains("2 Total"))
             .stdout(contains("2 OK"));
-
-        Ok(())
     }
 
     #[tokio::test]
     async fn test_cookie_jar() -> Result<()> {
         // Create a random cookie jar file
         let cookie_jar = NamedTempFile::new()?;
-        let mut cmd = main_command();
-        cmd.arg("--cookie-jar")
+        main_command!()
+            .arg("--cookie-jar")
             .arg(cookie_jar.path().to_str().unwrap())
             .arg("-")
             // Using Google as a test target because I couldn't
@@ -1810,57 +1868,51 @@ mod cli {
     }
 
     #[test]
-    fn test_dump_inputs_does_not_include_duplicates() -> Result<()> {
-        let pattern = fixtures_path().join("dump_inputs/markdown.md");
+    fn test_dump_inputs_does_not_include_duplicates() {
+        let pattern = fixtures_path!().join("dump_inputs/markdown.md");
 
-        let mut cmd = main_command();
-        cmd.arg("--dump-inputs")
+        main_command!()
+            .arg("--dump-inputs")
             .arg(&pattern)
             .arg(&pattern)
             .assert()
             .success()
             .stdout(contains("fixtures/dump_inputs/markdown.md").count(1));
-
-        Ok(())
     }
 
     #[test]
-    fn test_dump_inputs_glob_does_not_include_duplicates() -> Result<()> {
-        let pattern1 = fixtures_path().join("**/markdown.*");
-        let pattern2 = fixtures_path().join("**/*.md");
+    fn test_dump_inputs_glob_does_not_include_duplicates() {
+        let pattern1 = fixtures_path!().join("**/markdown.*");
+        let pattern2 = fixtures_path!().join("**/*.md");
 
-        let mut cmd = main_command();
-        cmd.arg("--dump-inputs")
+        main_command!()
+            .arg("--dump-inputs")
             .arg(pattern1)
             .arg(pattern2)
             .assert()
             .success()
             .stdout(contains("fixtures/dump_inputs/markdown.md").count(1));
-
-        Ok(())
     }
 
     #[test]
-    fn test_dump_inputs_glob_md() -> Result<()> {
-        let pattern = fixtures_path().join("**/*.md");
+    fn test_dump_inputs_glob_md() {
+        let pattern = fixtures_path!().join("**/*.md");
 
-        let mut cmd = main_command();
-        cmd.arg("--dump-inputs")
+        main_command!()
+            .arg("--dump-inputs")
             .arg(pattern)
             .assert()
             .success()
             .stdout(contains("fixtures/dump_inputs/subfolder/file2.md"))
             .stdout(contains("fixtures/dump_inputs/markdown.md"));
-
-        Ok(())
     }
 
     #[test]
-    fn test_dump_inputs_glob_all() -> Result<()> {
-        let pattern = fixtures_path().join("**/*");
+    fn test_dump_inputs_glob_all() {
+        let pattern = fixtures_path!().join("**/*");
 
-        let mut cmd = main_command();
-        cmd.arg("--dump-inputs")
+        main_command!()
+            .arg("--dump-inputs")
             .arg(pattern)
             .assert()
             .success()
@@ -1869,51 +1921,44 @@ mod cli {
             .stdout(contains("fixtures/dump_inputs/subfolder"))
             .stdout(contains("fixtures/dump_inputs/markdown.md"))
             .stdout(contains("fixtures/dump_inputs/some_file.txt"));
-
-        Ok(())
     }
 
     #[test]
-    fn test_dump_inputs_glob_exclude_path() -> Result<()> {
-        let pattern = fixtures_path().join("**/*");
+    fn test_dump_inputs_glob_exclude_path() {
+        let pattern = fixtures_path!().join("**/*");
 
-        let mut cmd = main_command();
-        cmd.arg("--dump-inputs")
+        main_command!()
+            .arg("--dump-inputs")
             .arg(pattern)
             .arg("--exclude-path")
-            .arg(fixtures_path().join("dump_inputs/subfolder"))
+            .arg(fixtures_path!().join("dump_inputs/subfolder"))
             .assert()
             .success()
             .stdout(contains("fixtures/dump_inputs/subfolder/test.html").not())
             .stdout(contains("fixtures/dump_inputs/subfolder/file2.md").not())
             .stdout(contains("fixtures/dump_inputs/subfolder").not());
-
-        Ok(())
     }
 
     #[test]
-    fn test_dump_inputs_url() -> Result<()> {
-        let mut cmd = main_command();
-        let result = cmd
+    fn test_dump_inputs_url() {
+        let result = main_command!()
             .arg("--dump-inputs")
             .arg("https://example.com")
             .assert()
             .success();
 
         assert_lines_eq(result, vec!["https://example.com/"]);
-        Ok(())
     }
 
     #[test]
-    fn test_dump_inputs_path() -> Result<()> {
-        let mut cmd = main_command();
-        let result = cmd
+    fn test_dump_inputs_path() {
+        let result = main_command!()
             .arg("--dump-inputs")
-            .arg(fixtures_path().join("dump_inputs"))
+            .arg(fixtures_path!().join("dump_inputs"))
             .assert()
             .success();
 
-        let base_path = fixtures_path().join("dump_inputs");
+        let base_path = fixtures_path!().join("dump_inputs");
         let expected_lines = [
             "some_file.txt",
             "subfolder/file2.md",
@@ -1925,17 +1970,15 @@ mod cli {
         .collect();
 
         assert_lines_eq(result, expected_lines);
-        Ok(())
     }
 
     // Ensures that dumping stdin does not panic and results in an empty output
     // as `stdin` is not a path
     #[test]
-    fn test_dump_inputs_with_extensions() -> Result<()> {
-        let mut cmd = main_command();
-        let test_dir = fixtures_path().join("dump_inputs");
+    fn test_dump_inputs_with_extensions() {
+        let test_dir = fixtures_path!().join("dump_inputs");
 
-        let output = cmd
+        let output = main_command!()
             .arg("--dump-inputs")
             .arg("--extensions")
             .arg("md,txt")
@@ -1952,7 +1995,7 @@ mod cli {
             .collect();
         actual_lines.sort();
 
-        let base_path = fixtures_path().join("dump_inputs");
+        let base_path = fixtures_path!().join("dump_inputs");
         let mut expected_lines = vec![
             path_str(&base_path, "some_file.txt"),
             path_str(&base_path, "subfolder/file2.md"),
@@ -1969,16 +2012,14 @@ mod cli {
                 "Should not contain example.bin: {line}"
             );
         }
-
-        Ok(())
     }
 
     #[test]
-    fn test_dump_inputs_skip_hidden() -> Result<()> {
-        let test_dir = fixtures_path().join("hidden");
+    fn test_dump_inputs_skip_hidden() {
+        let test_dir = fixtures_path!().join("hidden");
 
         // Test default behavior (skip hidden)
-        main_command()
+        main_command!()
             .arg("--dump-inputs")
             .arg(&test_dir)
             .assert()
@@ -1986,50 +2027,43 @@ mod cli {
             .stdout(is_empty());
 
         // Test with --hidden flag
-        main_command()
+        main_command!()
             .arg("--dump-inputs")
             .arg("--hidden")
             .arg(test_dir)
             .assert()
             .success()
             .stdout(contains(".hidden/file.md"));
-
-        Ok(())
     }
 
     #[test]
-    fn test_dump_inputs_individual_file() -> Result<()> {
-        let mut cmd = main_command();
-        let test_file = fixtures_path().join("TEST.md");
+    fn test_dump_inputs_individual_file() {
+        let test_file = fixtures_path!().join("TEST.md");
 
-        cmd.arg("--dump-inputs")
+        main_command!()
+            .arg("--dump-inputs")
             .arg(&test_file)
             .assert()
             .success()
             .stdout(contains("fixtures/TEST.md"));
-
-        Ok(())
     }
 
     #[test]
-    fn test_dump_inputs_stdin() -> Result<()> {
-        let mut cmd = main_command();
-
-        cmd.arg("--dump-inputs")
+    fn test_dump_inputs_stdin() {
+        main_command!()
+            .arg("--dump-inputs")
             .arg("-")
             .assert()
             .success()
             .stdout(contains("<stdin>"));
-
-        Ok(())
     }
 
     #[test]
     fn test_fragments_regression() {
-        let mut cmd = main_command();
-        let input = fixtures_path().join("FRAGMENT_REGRESSION.md");
+        let input = fixtures_path!().join("FRAGMENT_REGRESSION.md");
 
-        cmd.arg("--include-fragments")
+        main_command!()
+            .arg("--include-fragments")
             .arg("--verbose")
             .arg(input)
             .assert()
@@ -2038,10 +2072,9 @@ mod cli {
 
     #[test]
     fn test_fragments() {
-        let mut cmd = main_command();
-        let input = fixtures_path().join("fragments");
+        let input = fixtures_path!().join("fragments");
 
-        let mut result = cmd
+        let mut result = main_command!()
             .arg("--include-fragments")
             .arg("--verbose")
             .arg(input)
@@ -2131,12 +2164,12 @@ mod cli {
 
     #[test]
     fn test_fragments_when_accept_error_status_codes() {
-        let mut cmd = main_command();
-        let input = fixtures_path().join("TEST_FRAGMENT_ERR_CODE.md");
+        let input = fixtures_path!().join("TEST_FRAGMENT_ERR_CODE.md");
 
         // it's common for user to accept 429, but let's test with 404 since
         // triggering 429 may annoy the server
-        cmd.arg("--verbose")
+        main_command!()
+            .arg("--verbose")
             .arg("--accept=200,404")
             .arg("--include-fragments")
             .arg(input)
@@ -2152,10 +2185,10 @@ mod cli {
 
     #[test]
     fn test_fallback_extensions() {
-        let mut cmd = main_command();
-        let input = fixtures_path().join("fallback-extensions");
+        let input = fixtures_path!().join("fallback-extensions");
 
-        cmd.arg("--verbose")
+        main_command!()
+            .arg("--verbose")
             .arg("--fallback-extensions=htm,html")
             .arg(input)
             .assert()
@@ -2165,10 +2198,10 @@ mod cli {
 
     #[test]
     fn test_fragments_fallback_extensions() {
-        let mut cmd = main_command();
-        let input = fixtures_path().join("fragments-fallback-extensions");
+        let input = fixtures_path!().join("fragments-fallback-extensions");
 
-        cmd.arg("--include-fragments")
+        main_command!()
+            .arg("--include-fragments")
             .arg("--fallback-extensions=html")
             .arg("--no-progress")
             .arg("--offline")
@@ -2196,7 +2229,7 @@ mod cli {
     /// Note that the relative path is not resolved to the root of the server
     /// but relative to the file that contains the link.
     #[tokio::test]
-    async fn test_resolve_relative_paths_in_subfolder() -> Result<()> {
+    async fn test_resolve_relative_paths_in_subfolder() {
         let mock_server = wiremock::MockServer::start().await;
 
         let body = r#"<a href="next.html">next</a>"#;
@@ -2212,23 +2245,20 @@ mod cli {
             .mount(&mock_server)
             .await;
 
-        let mut cmd = main_command();
-        cmd.arg("--verbose")
+        main_command!()
+            .arg("--verbose")
             .arg(format!("{}/test/index.html", mock_server.uri()))
             .assert()
             .success()
             .stdout(contains("1 Total"))
             .stdout(contains("0 Errors"));
-
-        Ok(())
     }
 
     #[tokio::test]
     async fn test_json_format_in_config() -> Result<()> {
         let mock_server = mock_server!(StatusCode::OK);
-        let config = fixtures_path().join("configs").join("format.toml");
-        let mut cmd = main_command();
-        let output = cmd
+        let config = fixtures_path!().join("configs").join("format.toml");
+        let output = main_command!()
             .arg("--config")
             .arg(config)
             .arg("-")
@@ -2237,11 +2267,10 @@ mod cli {
             .assert()
             .success()
             .get_output()
-            .clone()
-            .unwrap();
+            .clone();
 
         // Check that the output is in JSON format
-        let output = std::str::from_utf8(&output.stdout).unwrap();
+        let output = std::str::from_utf8(&output.stdout)?;
         let json: serde_json::Value = serde_json::from_str(output)?;
         assert_eq!(json["total"], 1);
 
@@ -2252,8 +2281,7 @@ mod cli {
     async fn test_redirect_json() {
         use serde_json::json;
         redirecting_mock_server!(async |redirect_url: Url, ok_url| {
-            let mut cmd = main_command();
-            let output = cmd
+            let output = main_command!()
                 .arg("-")
                 .arg("--format")
                 .arg("json")
@@ -2287,7 +2315,7 @@ mod cli {
     }
 
     #[tokio::test]
-    async fn test_retry() -> Result<()> {
+    async fn test_retry() {
         let mock_server = wiremock::MockServer::start().await;
 
         wiremock::Mock::given(wiremock::matchers::method("GET"))
@@ -2301,18 +2329,15 @@ mod cli {
             .mount(&mock_server)
             .await;
 
-        let mut cmd = main_command();
-        cmd.arg("-")
+        main_command!()
+            .arg("-")
             .write_stdin(mock_server.uri())
             .assert()
             .success();
-
-        Ok(())
     }
 
     #[tokio::test]
-    async fn test_no_header_set_on_input() -> Result<()> {
-        let mut cmd = main_command();
+    async fn test_no_header_set_on_input() {
         let server = wiremock::MockServer::start().await;
         server
             .register(
@@ -2322,7 +2347,11 @@ mod cli {
             )
             .await;
 
-        cmd.arg("--verbose").arg(server.uri()).assert().success();
+        main_command!()
+            .arg("--verbose")
+            .arg(server.uri())
+            .assert()
+            .success();
 
         let received_requests = server.received_requests().await.unwrap();
         assert_eq!(received_requests.len(), 1);
@@ -2333,12 +2362,10 @@ mod cli {
 
         // Make sure the request does not contain the custom header
         assert!(!received_request.headers.contains_key("X-Foo"));
-        Ok(())
     }
 
     #[tokio::test]
-    async fn test_header_set_on_input() -> Result<()> {
-        let mut cmd = main_command();
+    async fn test_header_set_on_input() {
         let server = wiremock::MockServer::start().await;
         server
             .register(
@@ -2351,7 +2378,8 @@ mod cli {
             )
             .await;
 
-        cmd.arg("--verbose")
+        main_command!()
+            .arg("--verbose")
             .arg("--header")
             .arg("X-Foo: Bar")
             .arg(server.uri())
@@ -2360,12 +2388,10 @@ mod cli {
 
         // Check that the server received the request with the header
         server.verify().await;
-        Ok(())
     }
 
     #[tokio::test]
-    async fn test_multi_header_set_on_input() -> Result<()> {
-        let mut cmd = main_command();
+    async fn test_multi_header_set_on_input() {
         let server = wiremock::MockServer::start().await;
         server
             .register(
@@ -2379,7 +2405,8 @@ mod cli {
             )
             .await;
 
-        cmd.arg("--verbose")
+        main_command!()
+            .arg("--verbose")
             .arg("--header")
             .arg("X-Foo: Bar")
             .arg("--header")
@@ -2390,12 +2417,10 @@ mod cli {
 
         // Check that the server received the request with the header
         server.verify().await;
-        Ok(())
     }
 
     #[tokio::test]
-    async fn test_header_set_in_config() -> Result<()> {
-        let mut cmd = main_command();
+    async fn test_header_set_in_config() {
         let server = wiremock::MockServer::start().await;
         server
             .register(
@@ -2409,8 +2434,9 @@ mod cli {
             )
             .await;
 
-        let config = fixtures_path().join("configs").join("headers.toml");
-        cmd.arg("--verbose")
+        let config = fixtures_path!().join("configs").join("headers.toml");
+        main_command!()
+            .arg("--verbose")
             .arg("--config")
             .arg(config)
             .arg(server.uri())
@@ -2419,24 +2445,22 @@ mod cli {
 
         // Check that the server received the request with the header
         server.verify().await;
-        Ok(())
     }
 
     #[test]
-    fn test_sorted_error_output() -> Result<()> {
+    fn test_sorted_error_output() {
         let test_files = ["TEST_GITHUB_404.md", "TEST_INVALID_URLS.html"];
-
         let test_urls = [
             "https://httpbin.org/status/404",
             "https://httpbin.org/status/500",
             "https://httpbin.org/status/502",
         ];
 
-        let cmd = &mut main_command()
+        let cmd = &mut main_command!()
             .arg("--format")
             .arg("compact")
-            .arg(fixtures_path().join(test_files[1]))
-            .arg(fixtures_path().join(test_files[0]))
+            .arg(fixtures_path!().join(test_files[1]))
+            .arg(fixtures_path!().join(test_files[0]))
             .assert()
             .failure()
             .code(2);
@@ -2465,16 +2489,14 @@ mod cli {
             assert!(next_position > position);
             position = next_position;
         }
-
-        Ok(())
     }
 
     #[test]
     fn test_extract_url_ending_with_period_file() {
-        let test_path = fixtures_path().join("LINK_PERIOD.html");
+        let test_path = fixtures_path!().join("LINK_PERIOD.html");
 
-        let mut cmd = main_command();
-        cmd.arg("--dump")
+        main_command!()
+            .arg("--dump")
             .arg(test_path)
             .assert()
             .success()
@@ -2483,11 +2505,11 @@ mod cli {
 
     #[tokio::test]
     async fn test_extract_url_ending_with_period_webserver() {
-        let mut cmd = main_command();
         let body = r#"<a href="https://www.example.com/smth.">link</a>"#;
         let mock_server = mock_response!(body);
 
-        cmd.arg("--dump")
+        main_command!()
+            .arg("--dump")
             .arg(mock_server.uri())
             .assert()
             .success()
@@ -2496,10 +2518,10 @@ mod cli {
 
     #[test]
     fn test_wikilink_extract_when_specified() {
-        let test_path = fixtures_path().join("TEST_WIKI.md");
+        let test_path = fixtures_path!().join("TEST_WIKI.md");
 
-        let mut cmd = main_command();
-        cmd.arg("--dump")
+        main_command!()
+            .arg("--dump")
             .arg("--include-wikilinks")
             .arg(test_path)
             .assert()
@@ -2509,10 +2531,10 @@ mod cli {
 
     #[test]
     fn test_wikilink_dont_extract_when_not_specified() {
-        let test_path = fixtures_path().join("TEST_WIKI.md");
+        let test_path = fixtures_path!().join("TEST_WIKI.md");
 
-        let mut cmd = main_command();
-        cmd.arg("--dump")
+        main_command!()
+            .arg("--dump")
             .arg(test_path)
             .assert()
             .success()
@@ -2521,10 +2543,10 @@ mod cli {
 
     #[test]
     fn test_index_files_default() {
-        let input = fixtures_path().join("filechecker/dir_links.md");
+        let input = fixtures_path!().join("filechecker/dir_links.md");
 
         // the dir links in this file all exist.
-        main_command()
+        main_command!()
             .arg(&input)
             .arg("--verbose")
             .assert()
@@ -2533,7 +2555,7 @@ mod cli {
         // ... but checking fragments will find none, because dirs
         // have no fragments and no index file given.
         let dir_links_with_fragment = 2;
-        main_command()
+        main_command!()
             .arg(&input)
             .arg("--include-fragments")
             .assert()
@@ -2544,11 +2566,11 @@ mod cli {
 
     #[test]
     fn test_index_files_specified() {
-        let input = fixtures_path().join("filechecker/dir_links.md");
+        let input = fixtures_path!().join("filechecker/dir_links.md");
 
         // passing `--index-files index.html,index.htm` should reject all links
         // to /empty_dir because it doesn't have the index file
-        let result = main_command()
+        let result = main_command!()
             .arg(&input)
             .arg("--index-files")
             .arg("index.html,index.htm")
@@ -2566,7 +2588,7 @@ mod cli {
 
         // within the error message, formatting of the index file name list should
         // omit empty names.
-        main_command()
+        main_command!()
             .arg(&input)
             .arg("--index-files")
             .arg(",index.html,,,index.htm,")
@@ -2577,11 +2599,11 @@ mod cli {
 
     #[test]
     fn test_index_files_dot_in_list() {
-        let input = fixtures_path().join("filechecker/dir_links.md");
+        let input = fixtures_path!().join("filechecker/dir_links.md");
 
         // passing `.` in the index files list should accept a directory
         // even if no other index file is found.
-        main_command()
+        main_command!()
             .arg(&input)
             .arg("--index-files")
             .arg("index.html,.")
@@ -2592,7 +2614,7 @@ mod cli {
         // checking fragments will accept the index_dir#fragment link,
         // but reject empty_dir#fragment because empty_dir doesn’t have
         // index.html.
-        main_command()
+        main_command!()
             .arg(&input)
             .arg("--index-files")
             .arg("index.html,.")
@@ -2607,11 +2629,11 @@ mod cli {
 
     #[test]
     fn test_index_files_empty_list() {
-        let input = fixtures_path().join("filechecker/dir_links.md");
+        let input = fixtures_path!().join("filechecker/dir_links.md");
 
         // passing an empty list to --index-files should reject /all/
         // directory links.
-        let result = main_command()
+        let result = main_command!()
             .arg(&input)
             .arg("--index-files")
             .arg("")
@@ -2625,7 +2647,7 @@ mod cli {
             .stdout(contains("0 OK"));
 
         // ... as should passing a number of empty index file names
-        main_command()
+        main_command!()
             .arg(&input)
             .arg("--index-files")
             .arg(",,,,,")
@@ -2638,11 +2660,10 @@ mod cli {
     #[test]
     fn test_skip_binary_input() {
         // A path containing a binary file
-        let inputs = fixtures_path().join("invalid_utf8");
+        let inputs = fixtures_path!().join("invalid_utf8");
 
         // Run the command with the binary input
-        let mut cmd = main_command();
-        let result = cmd
+        let result = main_command!()
             .arg("--verbose")
             .arg(&inputs)
             .assert()
@@ -2668,11 +2689,11 @@ mod cli {
     #[test]
     fn test_dump_invalid_utf8_inputs() {
         // A path containing a binary file
-        let inputs = fixtures_path().join("invalid_utf8");
+        let inputs = fixtures_path!().join("invalid_utf8");
 
         // Run the command with the binary input
-        let mut cmd = main_command();
-        cmd.arg("--dump-inputs")
+        main_command!()
+            .arg("--dump-inputs")
             .arg(inputs)
             .assert()
             .success()
@@ -2687,7 +2708,7 @@ mod cli {
     /// See https://github.com/lycheeverse/lychee-action/issues/305
     #[test]
     fn test_globbed_files_are_always_checked() {
-        let input = fixtures_path().join("glob_dir/**/*.tsx");
+        let input = fixtures_path!().join("glob_dir/**/*.tsx");
 
         // The directory contains:
         // - example.ts
@@ -2695,7 +2716,7 @@ mod cli {
         // - example.md
         // - example.html
         // But the user only specified the .tsx file via the glob pattern.
-        main_command()
+        main_command!()
             .arg("--verbose")
             // Only check ts, js, and html files by default.
             // However, all files explicitly specified by the user
@@ -2710,11 +2731,11 @@ mod cli {
 
     #[test]
     fn test_extensions_work_on_glob_files_directory() {
-        let input = fixtures_path().join("glob_dir");
+        let input = fixtures_path!().join("glob_dir");
 
         // Make sure all files matching the given extensions are checked
         // if we specify a directory (and not a glob pattern).
-        main_command()
+        main_command!()
             .arg("--verbose")
             .arg("--extensions=ts,html")
             .arg(input)
@@ -2735,10 +2756,10 @@ mod cli {
     /// The extensions should only apply to the directory path, not the glob pattern.
     #[test]
     fn test_extensions_apply_to_files_not_globs() {
-        let glob_input = fixtures_path().join("glob_dir/**/*.tsx");
-        let dir_input = fixtures_path().join("example_dir");
+        let glob_input = fixtures_path!().join("glob_dir/**/*.tsx");
+        let dir_input = fixtures_path!().join("example_dir");
 
-        main_command()
+        main_command!()
             .arg("--verbose")
             .arg("--extensions=html,md")
             .arg(glob_input)
@@ -2756,7 +2777,7 @@ mod cli {
             .stderr(contains("https://example.com/example_dir/md"))
             // TS files in example_dir are ignored because we did not specify that extension.
             .stderr(contains("https://example.com/example_dir/ts ").not())
-            // TSX files in examle_dir are ignored because we did not specify that extension.
+            // TSX files in example_dir are ignored because we did not specify that extension.
             .stderr(contains("https://example.com/example_dir/tsx").not());
     }
 
@@ -2764,10 +2785,10 @@ mod cli {
     /// extension does not match the given extensions.
     #[test]
     fn test_file_inputs_always_get_checked_no_matter_their_extension() {
-        let ts_input_file = fixtures_path().join("glob_dir/example.ts");
-        let md_input_file = fixtures_path().join("glob_dir/example.md");
+        let ts_input_file = fixtures_path!().join("glob_dir/example.ts");
+        let md_input_file = fixtures_path!().join("glob_dir/example.md");
 
-        main_command()
+        main_command!()
             .arg("--verbose")
             .arg("--dump")
             .arg("--extensions=html,md")
@@ -2787,7 +2808,7 @@ mod cli {
     fn test_url_inputs_always_get_checked_no_matter_their_extension() {
         let url_input = "https://example.com/sitemap.xml";
 
-        main_command()
+        main_command!()
             .arg("--verbose")
             .arg("--dump")
             .arg(url_input)
@@ -2807,8 +2828,8 @@ mod cli {
         fs::write(&test_md, "# Test\n[link](https://example.com)")?;
         fs::write(&files_list_path, test_md.to_string_lossy().as_ref())?;
 
-        let mut cmd = main_command();
-        cmd.arg("--files-from")
+        main_command!()
+            .arg("--files-from")
             .arg(&files_list_path)
             .arg("--dump-inputs")
             .assert()
@@ -2826,8 +2847,8 @@ mod cli {
         // Create test file
         fs::write(&test_md, "# Test\n[link](https://example.com)")?;
 
-        let mut cmd = main_command();
-        cmd.arg("--files-from")
+        main_command!()
+            .arg("--files-from")
             .arg("-")
             .arg("--dump-inputs")
             .write_stdin(test_md.to_string_lossy().as_ref())
@@ -2854,8 +2875,8 @@ mod cli {
             ),
         )?;
 
-        let mut cmd = main_command();
-        cmd.arg("--files-from")
+        main_command!()
+            .arg("--files-from")
             .arg(&files_list_path)
             .arg("--dump-inputs")
             .assert()
@@ -2877,7 +2898,7 @@ mod cli {
         fs::write(&test_md2, "# Test 2")?;
         fs::write(&files_list_path, test_md1.to_string_lossy().as_ref())?;
 
-        let mut cmd = main_command();
+        let mut cmd = main_command!();
         cmd.arg("--files-from")
             .arg(&files_list_path)
             .arg(&test_md2) // Regular input argument
@@ -2891,16 +2912,14 @@ mod cli {
     }
 
     #[test]
-    fn test_files_from_nonexistent_file_error() -> Result<()> {
-        let mut cmd = main_command();
-        cmd.arg("--files-from")
+    fn test_files_from_nonexistent_file_error() {
+        main_command!()
+            .arg("--files-from")
             .arg("/nonexistent/file.txt")
             .arg("--dump-inputs")
             .assert()
             .failure()
             .stderr(contains("Cannot open --files-from file"));
-
-        Ok(())
     }
 
     /// Test the --default-extension option for files without extensions
@@ -2913,7 +2932,7 @@ mod cli {
         writeln!(file_without_ext, "[Local](local.md)")?;
 
         // Test with --default-extension md
-        main_command()
+        main_command!()
             .arg("--default-extension")
             .arg("md")
             .arg("--dump")
@@ -2932,7 +2951,7 @@ mod cli {
         writeln!(html_file_without_ext, "</body></html>")?;
 
         // Test with --default-extension html
-        main_command()
+        main_command!()
             .arg("--default-extension")
             .arg("html")
             .arg("--dump")
@@ -2954,7 +2973,7 @@ mod cli {
 
         // Unknown extensions should fall back to default behavior (plaintext)
         // and still extract links from the content
-        main_command()
+        main_command!()
             .arg("--default-extension")
             .arg("unknown")
             .arg("--dump")
@@ -2962,5 +2981,100 @@ mod cli {
             .assert()
             .success()
             .stdout(contains("https://example.org")); // Should extract the link as plaintext
+    }
+
+    /// An input which matches nothing should print a warning and continue.
+    #[test]
+    fn test_input_matching_nothing_warns() -> Result<()> {
+        let empty_dir = tempdir()?;
+
+        main_command!()
+            .arg(format!("{}", empty_dir.path().to_string_lossy()))
+            .arg(format!("{}/*", empty_dir.path().to_string_lossy()))
+            .arg("non-existing-path/*")
+            .arg("*.non-existing-extension")
+            .arg("non-existing-file-name???")
+            .assert()
+            .success()
+            .stderr(contains("No files found").count(5));
+
+        Ok(())
+    }
+
+    /// Invalid glob patterns should be checked and reported as a CLI parsing
+    /// error before link checking.
+    #[test]
+    fn test_invalid_glob_fails_parse() {
+        main_command!()
+            .arg("invalid-unmatched-brackets[")
+            .assert()
+            .stderr(contains("Cannot parse input"))
+            .failure()
+            .code(1); // cli parsing error code
+    }
+
+    /// Preprocessing with `cat` is like an identity function because it
+    /// outputs its input without any changes.
+    #[test]
+    fn test_pre_cat() {
+        let file = fixtures_path!().join("TEST.md");
+        let pre_with_cat = main_command!()
+            .arg("--preprocess")
+            .arg("cat")
+            .arg("--dump")
+            .arg(&file)
+            .assert()
+            .success();
+
+        let no_pre = main_command!()
+            .arg("--dump")
+            .arg(&file)
+            .assert()
+            .success()
+            .get_output()
+            .stdout
+            .lines()
+            .map(|line| line.unwrap().to_string())
+            .collect();
+
+        assert_lines_eq(pre_with_cat, no_pre);
+    }
+
+    #[test]
+    fn test_pre_invalid_command() {
+        let file = fixtures_path!().join("TEST.md");
+        main_command!()
+            .arg("--preprocess")
+            .arg("program does not exist")
+            .arg(file)
+            .assert()
+            .failure()
+            .stderr(contains("Error: Preprocessor command 'program does not exist' failed: could not start: No such file or directory (os error 2)"));
+    }
+
+    #[test]
+    fn test_pre_error() {
+        let file = fixtures_path!().join("TEST.md");
+        let script = fixtures_path!().join("pre").join("no_error_message.sh");
+        main_command!()
+            .arg("--preprocess")
+            .arg(&script)
+            .arg(&file)
+            .assert()
+            .failure()
+            .stderr(contains(format!(
+                "Error: Preprocessor command '{}' failed: exited with non-zero code: <empty stderr>", script.as_os_str().to_str().unwrap()
+            )));
+
+        let script = fixtures_path!().join("pre").join("error_message.sh");
+        main_command!()
+            .arg("--preprocess")
+            .arg(&script)
+            .arg(file)
+            .assert()
+            .failure()
+            .stderr(contains(format!(
+                "Error: Preprocessor command '{}' failed: exited with non-zero code: Some error message", script.as_os_str().to_str().unwrap()
+            )));
     }
 }
