@@ -1,24 +1,10 @@
 #[cfg(test)]
 mod readme {
-    use std::{fs, path::Path};
-
     use assert_cmd::Command;
     use pretty_assertions::assert_eq;
-
-    const USAGE_STRING: &str = "Usage: lychee [OPTIONS] [inputs]...\n";
-
-    fn main_command() -> Command {
-        // this gets the "main" binary name (e.g. `lychee`)
-        Command::cargo_bin(env!("CARGO_PKG_NAME")).expect("Couldn't get cargo package name")
-    }
-
-    fn load_readme_text() -> String {
-        let readme_path = Path::new(env!("CARGO_MANIFEST_DIR"))
-            .parent()
-            .unwrap()
-            .join("README.md");
-        fs::read_to_string(readme_path).unwrap()
-    }
+    use regex::Regex;
+    use test_utils::load_readme_text;
+    use test_utils::main_command;
 
     /// Remove line `[default: lychee/x.y.z]` from the string
     fn remove_lychee_version_line(string: &str) -> String {
@@ -44,20 +30,15 @@ mod readme {
     #[test]
     #[cfg(unix)]
     fn test_readme_usage_up_to_date() -> Result<(), Box<dyn std::error::Error>> {
-        let mut cmd = main_command();
+        const BEGIN: &str = "```help-message\n";
+        let mut cmd = main_command!();
 
         let help_cmd = cmd.env_clear().arg("--help").assert().success();
-        let help_output = std::str::from_utf8(&help_cmd.get_output().stdout)?;
-        let usage_in_help_start = help_output
-            .find(USAGE_STRING)
-            .ok_or("Usage not found in help")?;
-        let usage_in_help = &help_output[usage_in_help_start..];
+        let usage_in_help = std::str::from_utf8(&help_cmd.get_output().stdout)?;
 
         let usage_in_help = trim_empty_lines(&remove_lychee_version_line(usage_in_help));
-        let readme = load_readme_text();
-        let usage_start = readme
-            .find(USAGE_STRING)
-            .ok_or("Usage not found in README")?;
+        let readme = load_readme_text!();
+        let usage_start = readme.find(BEGIN).ok_or("Usage not found in README")? + BEGIN.len();
         let usage_end = readme[usage_start..]
             .find("\n```")
             .ok_or("End of usage not found in README")?;
@@ -65,6 +46,52 @@ mod readme {
         let usage_in_readme = remove_lychee_version_line(usage_in_readme);
 
         assert_eq!(usage_in_readme, usage_in_help);
+        Ok(())
+    }
+
+    /// Test that all the arguments yielded by `lychee --help`
+    /// are ordered alphabetically for better usability.
+    /// This behaviour aligns with cURL. (see `man curl`)
+    #[test]
+    #[cfg(unix)]
+    fn test_arguments_ordered_alphabetically() -> Result<(), Box<dyn std::error::Error>> {
+        let mut cmd = main_command!();
+        let help_cmd = cmd.env_clear().arg("--help").assert().success();
+        let help_text = std::str::from_utf8(&help_cmd.get_output().stdout)?;
+
+        let regex = test_utils::arg_regex_help!()?;
+        let arguments: Vec<&str> = help_text
+            .lines()
+            .filter_map(|line| {
+                let captures = regex.captures(line)?;
+                captures
+                    .name("short")
+                    .or_else(|| captures.name("long"))
+                    .map(|m| m.as_str())
+            })
+            .collect();
+
+        let mut sorted = arguments.clone();
+        sorted.sort_by_key(|arg| arg.to_lowercase());
+
+        if arguments != sorted {
+            // Find all positions where order differs
+            let mismatches: Vec<_> = arguments
+                .iter()
+                .zip(&sorted)
+                .enumerate()
+                .filter(|(_, (a, b))| a != b)
+                .map(|(i, (actual, expected))| format!("  [{i}] '{actual}' should be '{expected}'"))
+                .collect();
+
+            panic!(
+                "\nArguments are not sorted alphabetically!\n\nMismatches:\n{}\n\nFull actual order:\n{:?}\n\nFull expected order:\n{:?}",
+                mismatches.join("\n"),
+                arguments,
+                sorted
+            );
+        }
+
         Ok(())
     }
 }
