@@ -95,7 +95,7 @@ pub use crate::{
     types::{
         AcceptRange, AcceptRangeError, Base, BasicAuthCredentials, BasicAuthSelector, CacheStatus,
         CookieJar, ErrorKind, FileExtensions, FileType, Input, InputContent, InputResolver,
-        InputSource, Redirects, Request, ResolvedInputSource, Response, ResponseBody, Result,
-        Status, StatusCodeExcluder, StatusCodeSelector, uri::valid::Uri,
+        InputSource, Preprocessor, Redirects, Request, ResolvedInputSource, Response, ResponseBody,
+        Result, Status, StatusCodeExcluder, StatusCodeSelector, uri::valid::Uri,
     },
 };
