@@ -75,13 +75,13 @@ impl Remaps {
 
     /// Returns `true` if there is no remapping rule defined.
     #[must_use]
-    pub fn is_empty(&self) -> bool {
+    pub const fn is_empty(&self) -> bool {
         self.0.is_empty()
     }
 
     /// Get the number of remapping rules.
     #[must_use]
-    pub fn len(&self) -> usize {
+    pub const fn len(&self) -> usize {
         self.0.len()
     }
 }
