@@ -166,12 +166,11 @@ fn resolve_and_create_url(
 }
 
 fn prepend_root_dir_if_absolute_local_link(text: &str, root_dir: Option<&Path>) -> String {
-    if text.starts_with('/') {
-        if let Some(path) = root_dir {
-            if let Some(path_str) = path.to_str() {
-                return format!("{path_str}{text}");
-            }
-        }
+    if text.starts_with('/')
+        && let Some(path) = root_dir
+        && let Some(path_str) = path.to_str()
+    {
+        return format!("{path_str}{text}");
     }
     text.to_string()
 }
@@ -179,8 +178,23 @@ fn prepend_root_dir_if_absolute_local_link(text: &str, root_dir: Option<&Path>) 
 #[cfg(test)]
 mod tests {
     use std::borrow::Cow;
+    use std::num::NonZeroUsize;
+
+    use crate::types::uri::raw::RawUriSpan;
 
     use super::*;
+
+    fn raw_uri(text: &'static str) -> RawUri {
+        RawUri {
+            text: text.to_string(),
+            element: None,
+            attribute: None,
+            span: RawUriSpan {
+                line: NonZeroUsize::MAX,
+                column: None,
+            },
+        }
+    }
 
     #[test]
     fn test_is_anchor() {
@@ -200,7 +214,7 @@ mod tests {
         let base = Base::try_from("https://example.com/path/page.html").unwrap();
         let source = ResolvedInputSource::String(Cow::Borrowed(""));
 
-        let uris = vec![RawUri::from("relative.html")];
+        let uris = vec![raw_uri("relative.html")];
         let requests = create(uris, &source, None, Some(&base), None);
 
         assert_eq!(requests.len(), 1);
@@ -216,7 +230,7 @@ mod tests {
         let base = Base::try_from("https://example.com/path/page.html").unwrap();
         let source = ResolvedInputSource::String(Cow::Borrowed(""));
 
-        let uris = vec![RawUri::from("https://another.com/page")];
+        let uris = vec![raw_uri("https://another.com/page")];
         let requests = create(uris, &source, None, Some(&base), None);
 
         assert_eq!(requests.len(), 1);
@@ -232,7 +246,7 @@ mod tests {
         let base = Base::try_from("https://example.com/path/page.html").unwrap();
         let source = ResolvedInputSource::String(Cow::Borrowed(""));
 
-        let uris = vec![RawUri::from("/root-relative")];
+        let uris = vec![raw_uri("/root-relative")];
         let requests = create(uris, &source, None, Some(&base), None);
 
         assert_eq!(requests.len(), 1);
@@ -248,7 +262,7 @@ mod tests {
         let base = Base::try_from("https://example.com/path/page.html").unwrap();
         let source = ResolvedInputSource::String(Cow::Borrowed(""));
 
-        let uris = vec![RawUri::from("../parent")];
+        let uris = vec![raw_uri("../parent")];
         let requests = create(uris, &source, None, Some(&base), None);
 
         assert_eq!(requests.len(), 1);
@@ -264,7 +278,7 @@ mod tests {
         let base = Base::try_from("https://example.com/path/page.html").unwrap();
         let source = ResolvedInputSource::String(Cow::Borrowed(""));
 
-        let uris = vec![RawUri::from("#fragment")];
+        let uris = vec![raw_uri("#fragment")];
         let requests = create(uris, &source, None, Some(&base), None);
 
         assert_eq!(requests.len(), 1);
@@ -280,7 +294,7 @@ mod tests {
         let root_dir = PathBuf::from("/tmp/lychee");
         let source = ResolvedInputSource::FsPath(PathBuf::from("/some/page.html"));
 
-        let uris = vec![RawUri::from("relative.html")];
+        let uris = vec![raw_uri("relative.html")];
         let requests = create(uris, &source, Some((&root_dir, None)), None, None);
 
         assert_eq!(requests.len(), 1);
@@ -296,7 +310,7 @@ mod tests {
         let root_dir = PathBuf::from("/tmp/lychee");
         let source = ResolvedInputSource::FsPath(PathBuf::from("/some/page.html"));
 
-        let uris = vec![RawUri::from("https://another.com/page")];
+        let uris = vec![raw_uri("https://another.com/page")];
         let requests = create(uris, &source, Some((&root_dir, None)), None, None);
 
         assert_eq!(requests.len(), 1);
@@ -312,7 +326,7 @@ mod tests {
         let root_dir = PathBuf::from("/tmp/lychee");
         let source = ResolvedInputSource::FsPath(PathBuf::from("/tmp/lychee/page.html"));
 
-        let uris = vec![RawUri::from("/root-relative")];
+        let uris = vec![raw_uri("/root-relative")];
         let requests = create(uris, &source, Some((&root_dir, None)), None, None);
 
         assert_eq!(requests.len(), 1);
@@ -328,7 +342,7 @@ mod tests {
         let root_dir = PathBuf::from("/tmp/lychee");
         let source = ResolvedInputSource::FsPath(PathBuf::from("/some/page.html"));
 
-        let uris = vec![RawUri::from("../parent")];
+        let uris = vec![raw_uri("../parent")];
         let requests = create(uris, &source, Some((&root_dir, None)), None, None);
 
         assert_eq!(requests.len(), 1);
@@ -344,7 +358,7 @@ mod tests {
         let root_dir = PathBuf::from("/tmp/lychee");
         let source = ResolvedInputSource::FsPath(PathBuf::from("/some/page.html"));
 
-        let uris = vec![RawUri::from("#fragment")];
+        let uris = vec![raw_uri("#fragment")];
         let requests = create(uris, &source, Some((&root_dir, None)), None, None);
 
         assert_eq!(requests.len(), 1);
@@ -361,7 +375,7 @@ mod tests {
         let base = Base::try_from("https://example.com/path/page.html").unwrap();
         let source = ResolvedInputSource::FsPath(PathBuf::from("/tmp/lychee/localpage.html"));
 
-        let uris = vec![RawUri::from("relative.html")];
+        let uris = vec![raw_uri("relative.html")];
         let requests = create(uris, &source, Some((&root_dir, Some(&base))), None, None);
 
         assert_eq!(requests.len(), 1);
@@ -378,7 +392,7 @@ mod tests {
         let base = Base::try_from("https://example.com/path/page.html").unwrap();
         let source = ResolvedInputSource::FsPath(PathBuf::from("/some/page.html"));
 
-        let uris = vec![RawUri::from("https://another.com/page")];
+        let uris = vec![raw_uri("https://another.com/page")];
         let requests = create(uris, &source, Some((&root_dir, Some(&base))), None, None);
 
         println!("{:?}", requests);
@@ -396,7 +410,7 @@ mod tests {
         let base = Base::try_from("https://example.com/path/page.html").unwrap();
         let source = ResolvedInputSource::FsPath(PathBuf::from("/tmp/lychee/localpage.html"));
 
-        let uris = vec![RawUri::from("/root-relative")];
+        let uris = vec![raw_uri("/root-relative")];
         let requests = create(uris, &source, Some((&root_dir, Some(&base))), None, None);
 
         println!("{:?}", requests);
@@ -414,7 +428,7 @@ mod tests {
         let base = Base::try_from("https://example.com/path/page.html").unwrap();
         let source = ResolvedInputSource::FsPath(PathBuf::from("/some/page.html"));
 
-        let uris = vec![RawUri::from("../parent")];
+        let uris = vec![raw_uri("../parent")];
         let requests = create(uris, &source, Some((&root_dir, Some(&base))), None, None);
 
         assert_eq!(requests.len(), 1);
@@ -431,7 +445,7 @@ mod tests {
         let base = Base::try_from("https://example.com/path/").unwrap();
         let source = ResolvedInputSource::FsPath(PathBuf::from("/some/page.html"));
 
-        let uris = vec![RawUri::from("#fragment")];
+        let uris = vec![raw_uri("#fragment")];
         let requests = create(uris, &source, Some((&root_dir, Some(&base))), None, None);
 
         assert_eq!(requests.len(), 1);
@@ -446,7 +460,7 @@ mod tests {
     fn test_no_base_url_resolution() {
         let source = ResolvedInputSource::String(Cow::Borrowed(""));
 
-        let uris = vec![RawUri::from("https://example.com/page")];
+        let uris = vec![raw_uri("https://example.com/page")];
         let requests = create(uris, &source, None, None, None);
 
         assert_eq!(requests.len(), 1);
@@ -464,7 +478,7 @@ mod tests {
         let base_info = SourceBaseInfo::from_source(&input_source, None, Some(&base)).unwrap();
 
         let actual =
-            create_request(&RawUri::from("file.html"), &input_source, &base_info, None).unwrap();
+            create_request(&raw_uri("file.html"), &input_source, &base_info, None).unwrap();
 
         assert_eq!(
             actual,
@@ -488,7 +502,7 @@ mod tests {
 
         // Use an absolute path that's outside the base directory
         let actual = create_request(
-            &RawUri::from("/usr/local/share/doc/example.html"),
+            &raw_uri("/usr/local/share/doc/example.html"),
             &input_source,
             &base_info,
             None,
@@ -514,7 +528,7 @@ mod tests {
         let base = Base::Local(PathBuf::from("/tmp/lychee"));
         let source = ResolvedInputSource::String(Cow::Borrowed(""));
 
-        let raw_uri = RawUri::from("relative.html");
+        let raw_uri = raw_uri("relative.html");
         let uri = try_parse_into_uri(&raw_uri, &source, None, Some(&base)).unwrap();
 
         assert_eq!(uri.url.as_str(), "file:///tmp/lychee/relative.html");
@@ -525,7 +539,7 @@ mod tests {
         let base = Base::Local(PathBuf::from("/tmp/lychee"));
         let source = ResolvedInputSource::String(Cow::Borrowed(""));
 
-        let raw_uri = RawUri::from("absolute.html");
+        let raw_uri = raw_uri("absolute.html");
         let uri = try_parse_into_uri(&raw_uri, &source, None, Some(&base)).unwrap();
 
         assert_eq!(uri.url.as_str(), "file:///tmp/lychee/absolute.html");
