@@ -107,11 +107,18 @@ macro_rules! mail {
     }};
 }
 
-/// Returns the path to the `fixtures` directory.
-///
-/// # Panic
-///
-/// Panics if the fixtures directory could not be determined.
+/// Get the root path of the project.
+#[macro_export]
+macro_rules! root_path {
+    () => {
+        std::path::Path::new(env!("CARGO_MANIFEST_DIR"))
+            .parent()
+            .unwrap()
+            .to_path_buf()
+    };
+}
+
+/// Get the path to the `fixtures` directory.
 #[macro_export]
 macro_rules! fixtures_path {
     () => {
@@ -148,4 +155,42 @@ macro_rules! fixture_uri {
             .join($subpath)
             .expect("expected subpath to form a valid URL")
     }};
+}
+
+#[macro_export]
+macro_rules! load_readme_text {
+    () => {{
+        let readme_path = std::path::Path::new(env!("CARGO_MANIFEST_DIR"))
+            .parent()
+            .unwrap()
+            .join("README.md");
+        std::fs::read_to_string(readme_path).unwrap()
+    }};
+}
+
+/// Helper function to create a ResponseBody with a given status and URI
+#[macro_export]
+macro_rules! mock_response_body {
+    ($status:expr, $uri:expr $(,)?) => {{
+        ResponseBody {
+            uri: Uri::try_from($uri).unwrap(),
+            status: $status,
+        }
+    }};
+}
+
+/// Gets the "main" binary name (e.g. `lychee`)
+#[macro_export]
+macro_rules! main_command {
+    () => {
+        Command::cargo_bin(env!("CARGO_PKG_NAME")).expect("Couldn't get cargo package name")
+    };
+}
+
+/// Capture all CLI flags (e.g. `-a` or `--accept`) from the help message via regex
+#[macro_export]
+macro_rules! arg_regex_help {
+    () => {
+        Regex::new(r"^\s{2,6}(?:-(?<short>[a-zA-Z]),)?\s--(?<long>[a-zA-Z-]+)")
+    };
 }
